backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.15
mysqlclient==2.0.3
Werkzeug==2.0.1
flask-cors==3.0.10
mysql-connector-python